+++ src/editors/tools.js
@@ -1,5 +1,5 @@
 import { addSettlement, removeSettlement } from "../generators/civilization.js";
-import { MARKER_TYPES } from "../generators/markers.js";
+import { MARKER_TYPES } from "../data/catalogs.js";
 import { stampAt } from "../generators/landforms.js";
 import { addRouteBetween, removeRoutesThrough, pruneRoutes } from "../generators/routes.js";
 import { cellsInBrush, eyedrop, paintCells } from "./paint.js";